+++ documentcloud/documents/models/__init__.py
@@ -1,3 +1,4 @@
 # DocumentCloud
 from documentcloud.documents.models.document import *
 from documentcloud.documents.models.entity import *
+from documentcloud.documents.models.note import *
